feat(configurations): add ConfigurationService class
Introduces the `ConfigurationService` class in the `Void.Proxy.Configurations` namespace, implementing the `IConfigurationService` interface. The class includes a generic method `Get<TConfiguration>` that accepts an `IPlugin` parameter and returns a default value of type `TConfiguration`, constrained to implement the `IConfiguration` interface. New using directives for `NuGet.Protocol.Plugins` and `Void.Proxy.Api.Configurations` are also added to support this functionality.