feat(packet): add constructor for ChatMessagePacket v1.20.2-3
This commit introduces a new constructor for the `ChatMessagePacket` class that supports protocol versions from Minecraft 1.20.2 to 1.20.3. It includes a call to `ComponentJsonTransformers.Passthrough_v1_20_2_to_v1_20_3(wrapper)` and passes through `ByteProperty` and `UuidProperty`. A comment has also been added to indicate that this packet is ignored due to its removal after version 1.18.