docs: update README.md for proxy forwarding capabilities
Clarified the forwarding capabilities of the proxy in the README.md file.
Added information about playing through the proxy without a private key, noting that the protocol API will be limited in that case.
Removed a line suggesting that playing in online mode with Void is not recommended, which may present a more positive view of using the proxy in online mode.