feat: enhance service registration in HostingExtensions
Modified the `Add` method in the `HostingExtensions` class to delegate service registration to a new method that accepts an `IContainer` parameter. This change improves separation of concerns and enhances the overall structure of the code. A new method signature for the `Add` method that operates on `IContainer` has also been introduced.