docs: clarify Link purpose
Enhanced the readability of the `ChatService` class by adding comments that explain the concept of a "link" as a connection between a player and the server. The comments also note that a new link will be created if a player is redirected to another server. This improves context for future developers and maintainers.