fix: more crossplatform plugin devkit
Updated `launchSettings.json` to enable hot reload functionality. Modified the `workingDirectory` to point to the `$(TargetDir)void` directory.

In `YourPlugin.csproj`, added a new `Setup` target that defines the `<VoidPlatform>` and `<VoidExtension>`, removes previous platform definitions, and includes tasks to download the `void` executable and make it executable on non-Windows platforms.