fix(transformers): ✨ update hoverEvent processing to support Minecraft Console Client behavior
Modified the handling of "hoverEvent" actions in both
ComponentJsonTransformers and ComponentNbtTransformers to
check for the presence of a "text" key in the associated
value. This change accommodates Minecraft Console Clients
that may send a JsonObject or NbtCompound directly as the
value while preserving the original behavior for Vanilla
clients during the upgrade from version 1.8 to 1.9.