docs(routes): 🔄 update last modified date handling
The last modified date retrieval for routes has been improved by replacing the `routeLastmod` functionality with a new `getLatestCommit` function. This change enhances the clarity and maintainability of the codebase.

The sitemap serialization now utilizes the new function to set the `lastmod` property for items. Additionally, the `LastUpdated.astro` file has been updated to check for a `lastUpdated` property, improving the display of commit dates.

The `update-time.ts` file has been restructured to remove the old map and introduce a new one for storing commit information, streamlining the process of collecting and retrieving commit data.