docs(features): ✏️ restructure features presentation
Updated the `features.mdx` documentation to enhance the presentation of features supported by the Void proxy. Introduced badges for terms, improved the game versions table, and added a new section on forwarding methods. The mod loaders section was updated, and the focus has shifted to Minecraft, removing references to other games.